Progreso datos forecast

+++ src/main/java/com/ccsw/teammanager/groupmembers/model/GroupMembersDto.java
@@ -4,9 +4,9 @@ public class GroupMembersDto {
 
     private String id;
 
-    private Integer group_id;
+    private Long group_id;
 
-    private Integer person_id;
+    private Long person_id;
 
     public String getId() {
         return id;
@@ -16,19 +16,19 @@ public class GroupMembersDto {
         this.id = id;
     }
 
-    public Integer getGroup_id() {
+    public Long getGroup_id() {
         return group_id;
     }
 
-    public void setGroup_id(Integer group_id) {
+    public void setGroup_id(Long group_id) {
         this.group_id = group_id;
     }
 
-    public Integer getPerson_id() {
+    public Long getPerson_id() {
         return person_id;
     }
 
-    public void setPerson_id(Integer person_id) {
+    public void setPerson_id(Long person_id) {
         this.person_id = person_id;
     }
